Change type to float32

+++ pkg/api/http/requests/create_expense_summary.go
@@ -7,6 +7,6 @@ import (
 type CreateExpenseSummaryRequest struct {
 	Name                      string
 	ClosedAt                  time.Time
-	UsdToPlnRatio             float64
-	MoneyTransferredToSavings float64
+	UsdToPlnRatio             float32
+	MoneyTransferredToSavings float32
 }
