+++ pkg/api/http/responses/create_expenses_summary.go
@@ -11,6 +11,6 @@ type CreateExpensesSummaryResponse struct {
 	CreatedAt                 time.Time
 	UpdatedAt                 time.Time
 	ClosedAt                  *time.Time `json:",omitempty"`
-	UsdToPlnRatio             float64
-	MoneyTransferredToSavings float64
+	UsdToPlnRatio             float32
+	MoneyTransferredToSavings float32
 }
